Add support for PEP 604 and forward reference
Prior to this commit, using a forward reference type annotation
(i.e., a string) would fail to be detected properly. Similarly,
a PEP 604 annotation using the pipe operator would fail.

This commit adds support and tests for both.

Also, either W503 or W504 has to be ignored, so we ignore the
one that says will be an anti-pattern.

Also, the toml file still claimed we support Python3.7, but
we had no CI/CD support for it, so we updated the toml file.