feat: Add Dict[str, str] via logfmt
Prior to this commit, a value could be a list of string,
but not a dictionary of strings.

This commit resolves that by adding logfmt loader.

Also, this commit moves the tests around, since the
package is getting bigger, so we want a bit more
organisation. The loader tests are also more simplified,
so you don't need to set up a environ every time.

And since we're improving things anyhow, adds code
coverage generation.

Lastly, PP UP.